Added left walk animation and a "minimap"
I added a "rough" minimap, because I don't know what the function of it
is atm, should it display the whole map with the player in it on a small
scale on the top left/right corner or should it follow the player. Once
we decide on that, I will then make further changes to it.